refactor info into json object
