--- src/pages/BondDetail/bondInformation.tsx
@@ -1,21 +1,0 @@
-const bondInformation = [
-  {
-    bond: '0x11f1f978f7944579bb3791b765176de3e68bffc6',
-    prime: 'https://www.prime.xyz/ratings/shapeshift',
-    defiLlama: 'https://defillama.com/protocol/shapeshift',
-  },
-  {
-    bond: '0xe34c023c0ea9899a8f8e9381437a604908e8b719',
-    creditAnalysis: '/pdf/Ribbon DAO Collateral & Credit Analysis.pdf',
-    prime: 'https://www.prime.xyz/ratings/ribbon-finance',
-    defiLlama: 'https://defillama.com/protocol/ribbon',
-  },
-  {
-    bond: '0x0ce1f1cd784bd2341abf21444add0681fe5a526c',
-    auction: '20',
-    prime: 'https://www.prime.xyz/ratings/shapeshift',
-    defiLlama: 'https://defillama.com/protocol/shapeshift',
-  },
-]
-
-export { bondInformation }
